Adecuacion para la carga de la constancia INFONAVIT

+++ main.py
@@ -140,16 +140,20 @@ async def get_docs(rfc: str,req: str):
 
 @app.post("/infonavit/uploadopinion")
 async def upload_opinion(rfc:str,file:UploadFile = File(...)):
-    descarga = "/root/"+rfc+"/INFONAVIT"
-    #descarga = "E:\\SAT\\"+rfc+"\\INFONAVIT"    
-    folder_path = Path(descarga)
-    folder_path.mkdir(parents=True, exist_ok=True)
-    for file in pathlib.Path(descarga).glob('*.*'):
-        try:
-            file.unlink()
-        except:
-            pass
-    with open(f"{descarga}/{file.filename}", "wb") as buffer:
-        shutil.copyfileobj(file.file, buffer)
+    if file.content_type != "application/pdf":
+        descarga = "/root/"+rfc+"/INFONAVIT"
+        #descarga = "E:\\SAT\\"+rfc+"\\INFONAVIT"    
+        folder_path = Path(descarga)
+        folder_path.mkdir(parents=True, exist_ok=True)
+        for file in pathlib.Path(descarga).glob('*.*'):
+            try:
+                file.unlink()
+            except:
+                pass
+        with open(f"{descarga}/{file.filename}", "wb") as buffer:
+            shutil.copyfileobj(file.file, buffer)
+        return {"filename" : file.filename, "message" : "Archivo cargado satisfactoriamente", "Tipo archivo":file.content_type}
+    else
+        return {"filename" : file.filename, "message" : "El archivo no corresponde a un documento PDF", "Tipo archivo":file.content_type}
 
-    return {"filename" : file.filename, "message" : "Archivo cargado satisfactoriamente", "Tipo archivo":file.content_type}
+
